requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.4
tensorflow==2.12.0
Pillow==10.3.0
numpy==1.26.4
opencv-python==4.9.0.80
matplotlib==3.8.4
